Completely reworked the library search
The local library music search is now on its own thread.

The thread will run every so often depending on the config setting. It will
then update a list of all the music files in local memory. When a user searches
for music, it will serach this list in memory instead of going to the disk. This
results in much faster/near instant playback.