Trying to fix the issue with calling music command while music is playing
There is an exception for audio already playing when calling the music
command while music is playing. Attempting to fix that, however it is
still occuring. This change at least fixes the queue. Previously,
when calling music while music was already playing, it would then
skip the first song of the new queue.