Cogs music update (#13)
* Update to use Cogs

* Inital code for music with Cogs

* Allow for searching and playing back local albums

Fixed !stop so it correctly stops playback and clears the queue

Updated README.md config.json

* Added music functionality to go to previous songs

Added check to make sure it doesn't attempt to play an empty queue
after a search and if so it will send a message that no songs could be
found

* Fixed !next and !prev commands by reworking the queue

Display which song is currently playing each time the song changes

Only attempt to play certain audio files (no jpg, png, and so on)

* Load audio metadata and display details to the user about the current song

* Set currSong to 0 when stopping and clearing the queue

* Added after invocation hook for !next, and display the current song

This should force the function to now always display the correct
song since it will always run after !next is completed.

* Added some f strings

* Updated !prev command to display correct song metadata.

Fixed an issue in requirements

* Added !volume command

!volume accepts values between 0 and 100 and modifies the volume of the
music client.

* Completely reworked the library search

The local library music search is now on its own thread.

The thread will run every so often depending on the config setting. It will
then update a list of all the music files in local memory. When a user searches
for music, it will serach this list in memory instead of going to the disk. This
results in much faster/near instant playback.

* Fixed an exception which would occur when using !music

Using !music when not in a voice channel would cause an error since the
bot would not be able to join the voice channel. Now it responds with a
message to the user telling them to join a voice channel.

* Exception handling for the bot not having permissions to a channel

* Docstrings, queue, come, pause, resume and bug fixes

Added many docstrings which in turn populate the help command
Added queue command which displays the previous, current, and next songs
Added come command which moves the bot to your current voice channel
Added pause and resume commands
Fixed several bugs pertaining to index out of range errors

* Removed some f strings which weren't required.

* Removed some mentions

* Fix blank exceptions and removed an fstring

* Trying to fix the issue with calling music command while music is playing

There is an exception for audio already playing when calling the music
command while music is playing. Attempting to fix that, however it is
still occuring. This change at least fixes the queue. Previously,
when calling music while music was already playing, it would then
skip the first song of the new queue.

* Force music command to wait until inital local music search is completed

If a user uses the music command very quickly after the bot is first started
the local music library may not be loaded into memory just yet.
Put a lock into place so that the music command waits until the
inital search is completed.

* Load config into music from the config.yaml

Fix merge issues with threading branch

* Add required params for Music

* Fixed README for this branch update

* Removed double quotes on some readme values

* Fixed windows ctrl+c not exiting the program