Fixed timing of cli startup
The CLI was starting up whenever its thread was able to. This caused
some oddities when it would first start up as the rest of the bots
and cogs were not always ready. Added a ready variable to each cog
to be set when it is ready. The cli now listens for these ready signals
and will start allowing input in the cli once all cogs and bots are ready