Added permission checks to insults (#43)
* Added permission checks to insults

Permissions are now pulled from the config.yaml for each insult
command. There will be a check on each command to ensure the user
has access to use the command. If not, a message will be sent to
the user that they do not have permissions to use the command.

* Moved check_permissions

* Backport music permission changes to insult