Added shuffle command to music
The shuffle command will mix up the current music queue.

It takes the current playing song, saves it, shuffles the queue
then places the current song at the beginning of the queue.
It does all this without interrupting the current audio stream.

The result will be a new queue with the current song at the beginning.