Music cleanup (#50)
* Move check_permissions to TGACog parent class

* Init the TGACog parent class from within each Cog

* Moved load permissions to TGACog

* Moved CONFIG definition to TGACog init

Instead of defining the CONFIG inside the process_config method
define it in the init of TGACog. Also define cog_name in the init
as well. This cleans up some of the TGACog class methods and makes
them a bit easier to read.

* Renamed cog_name to COG_NAME

* Missed a cog_name rename

* Cleaned up music cog

Fixed issues with moving the bot to other channels introduced in
discord.py 1.5.0

Upgraded discord.py requirement to 1.5.1

Added new methods to check for voice channel connectivity and
if the user making a command is connected to a voice channel

Made a number of methods private

* Removed random log in code

Co-authored-by: Rob D <rcdelga@outlook.com>