Music shuffle (#52)
* Move check_permissions to TGACog parent class

* Init the TGACog parent class from within each Cog

* Moved load permissions to TGACog

* Moved CONFIG definition to TGACog init

Instead of defining the CONFIG inside the process_config method
define it in the init of TGACog. Also define cog_name in the init
as well. This cleans up some of the TGACog class methods and makes
them a bit easier to read.

* Renamed cog_name to COG_NAME

* Missed a cog_name rename

* Cleaned up music cog

Fixed issues with moving the bot to other channels introduced in
discord.py 1.5.0

Upgraded discord.py requirement to 1.5.1

Added new methods to check for voice channel connectivity and
if the user making a command is connected to a voice channel

Made a number of methods private

* Added shuffle command to music

The shuffle command will mix up the current music queue.

It takes the current playing song, saves it, shuffles the queue
then places the current song at the beginning of the queue.
It does all this without interrupting the current audio stream.

The result will be a new queue with the current song at the beginning.

* Added a message when music is shuffled or if there is no music to shuffle

Updated bot so that if there is a feature defined in the config.yaml
it doesn't raise an error and instead logs the error and continues.

* Removed fstring

* Removed a variable which wasn't needed

* Tiny refactor to shuffle

Co-authored-by: Rob D <rcdelga@outlook.com>